refactor: moved models to lib package
